+++ internal/table/scanner/scanner.go
@@ -414,7 +414,6 @@ func (s *valueScanner) any() interface{} {
 	case internalTypes.Bytes:
 		return s.bytes()
 	case internalTypes.UUID:
-		// replace to good uuid on migration
 		return s.uuidBytesWithIssue1501()
 	case internalTypes.Uint32:
 		return s.uint32()